requirements.txt
vllm==0.8.2
huggingface_hub
transformers
tqdm
wandb
ipywidgets
bitsandbytes
accelerate
datasets
hydra-core
omegaconf
evaluate
pydantic
trl==0.15.2
peft
deepspeed==0.15.4
cachetools
scikit-learn
evaluate
jinja2
colorama
flash_attn
multiprocess
pydantic
PyYAML
polars
numpy==1.25.2
